added dynamic country/flavor combination

+++ pd/pdcalc.py
@@ -6,6 +6,7 @@ import requests
 import os
 import sys
 from os import path
+import json
 
 import cacher
 
@@ -13,8 +14,8 @@ sys.path.insert(0, "/usr/lib/python2.7/dist-packages") #to have redland librdf l
 
 if __name__ == '__main__':
   parser = argparse.ArgumentParser(description='Public Domain Calculator')
-  parser.add_argument('-c', '--country', dest='country',  help='country for which to test', required=True)
-  parser.add_argument('-i', '--instance', dest='instance',  help='instance to test', required=True)
+  parser.add_argument('-c', '--country', dest='country',  help='country for which to test')
+  parser.add_argument('-i', '--instance', dest='instance',  help='instance to test')
   parser.add_argument('-l', '--flavor', dest='flavor',  help='specific flavor', default="")
   parser.add_argument('-g', '--global', dest='globalmap',  help='global mapping', default="global.json")
   parser.add_argument('-m', '--mode', dest='mode',  help='instance definition mode', choices=['file', 'url'], default="file")
@@ -24,38 +25,53 @@ if __name__ == '__main__':
   parser.add_argument('-q', '--query', dest="query", help="one-shot query", default=None)
   parser.add_argument('-L', '--language', dest="language", help="language", default="en")
 
+  parser.add_argument('-V', '--valid', dest="valid", help="valid country and flavors", default=False,action='store_true')
+
   args = parser.parse_args()
 
-  if args.instance.startswith("http"):
-    args.mode = "url"
+  if args.valid:
 
-  if args.mode == "url":
-    f = tempfile.NamedTemporaryFile("w", delete=False)
-    data = cacher.get(args.instance)
-    #print data.encoding
-    #print data.text
-    #if data.status_code != 200:
-    #  raise Exception("Wrong Status ==> Network Error.")
-    f.write(data)
-    f.close()
-    args.instance = f.name
-
-  files = [ os.path.join(args.country,"flow.json"), os.path.join(args.country,args.flavor, "local.json"), os.path.join(args.country, "local.json"), args.globalmap, args.instance]
-  #print files
-  files = map(path.isfile, files)
-  #print files
-  if all(files):
-    from reasoner import Reasoner
-    a = Reasoner(os.path.join(args.country,"flow.json"), local_map = os.path.join(args.country, "local.json"), flavor_map = os.path.join(args.country,args.flavor, "local.json"), global_map = args.globalmap, detail=args.detail, output=args.output, language=os.path.join(args.country, "i18n", args.language+".json"))
-    a.parse_input(args.instance)
-    a.info()
-    if args.query is None:
-      a.run()
+    itms = []
+    for x in os.walk('.'):
+      if not x[0].endswith("i18n") and not x[0] == ".":
+        itms.append(x[0][2:])
+    if args.output == "json":
+      print json.dumps(itms)
     else:
-      a.query(args.query)
-    print a.get_result()
+      print ", ".join(itms)
 
-  if args.mode == "url":
-    os.remove(args.instance)
+  else:
 
-  sys.exit()
+    if args.instance.startswith("http"):
+      args.mode = "url"
+
+    if args.mode == "url":
+      f = tempfile.NamedTemporaryFile("w", delete=False)
+      data = cacher.get(args.instance)
+      #print data.encoding
+      #print data.text
+      #if data.status_code != 200:
+      #  raise Exception("Wrong Status ==> Network Error.")
+      f.write(data)
+      f.close()
+      args.instance = f.name
+
+    files = [ os.path.join(args.country,"flow.json"), os.path.join(args.country,args.flavor, "local.json"), os.path.join(args.country, "local.json"), args.globalmap, args.instance]
+    #print files
+    files = map(path.isfile, files)
+    #print files
+    if all(files):
+      from reasoner import Reasoner
+      a = Reasoner(os.path.join(args.country,"flow.json"), local_map = os.path.join(args.country, "local.json"), flavor_map = os.path.join(args.country,args.flavor, "local.json"), global_map = args.globalmap, detail=args.detail, output=args.output, language=os.path.join(args.country, "i18n", args.language+".json"))
+      a.parse_input(args.instance)
+      a.info()
+      if args.query is None:
+        a.run()
+      else:
+        a.query(args.query)
+      print a.get_result()
+
+    if args.mode == "url":
+      os.remove(args.instance)
+
+    sys.exit()
